# [PIM-6670] Optimize attributes normalization