--- a/src/Pim/Bundle/EnrichBundle/Resources/config/controllers.yml
+++ b/src/Pim/Bundle/EnrichBundle/Resources/config/controllers.yml
@@ -202,6 +202,7 @@ services:
             - '@pim_catalog.factory.attribute'
             - '@pim_user.context.user'
             - '@pim_catalog.localization.localizer.number'
+            - '@pim_enrich.normalizer.attribute'
 
     pim_enrich.controller.rest.attribute_type:
         class: '%pim_enrich.controller.rest.attribute_type.class%'