--- a/src/Pim/Bundle/EnrichBundle/Resources/config/normalizers.yml
+++ b/src/Pim/Bundle/EnrichBundle/Resources/config/normalizers.yml
@@ -11,6 +11,7 @@ parameters:
     pim_enrich.normalizer.group_type.class:                        Pim\Bundle\EnrichBundle\Normalizer\GroupTypeNormalizer
     pim_enrich.normalizer.family.class:                            Pim\Bundle\EnrichBundle\Normalizer\FamilyNormalizer
     pim_enrich.normalizer.attribute.class:                         Pim\Bundle\EnrichBundle\Normalizer\AttributeNormalizer
+    pim_enrich.normalizer.attribute.versioned.class:               Pim\Bundle\EnrichBundle\Normalizer\VersionedAttributeNormalizer
     pim_enrich.normalizer.locale.class:                            Pim\Bundle\EnrichBundle\Normalizer\LocaleNormalizer
     pim_enrich.normalizer.channel.class:                           Pim\Bundle\EnrichBundle\Normalizer\ChannelNormalizer
     pim_enrich.normalizer.sequential_edit.class:                   Pim\Bundle\EnrichBundle\Normalizer\SequentialEditNormalizer
@@ -144,10 +145,15 @@ services:
             - '@pim_enrich.provider.field.chained'
             - '@pim_enrich.provider.empty_value.chained'
             - '@pim_enrich.provider.filter.chained'
+            - '@pim_catalog.localization.localizer.number'
+
+    pim_enrich.normalizer.attribute.versioned:
+        class: '%pim_enrich.normalizer.attribute.versioned.class%'
+        arguments:
+            - '@pim_enrich.normalizer.attribute'
             - '@pim_versioning.manager.version'
             - '@pim_enrich.normalizer.version'
             - '@pim_enrich.provider.structure_version.attribute'
-            - '@pim_catalog.localization.localizer.number'
         tags:
             - { name: pim_internal_api_serializer.normalizer }