**The context**
We need to load the PEF as fast as possible.

**The problem**
Among the most blocking requests : the one getting the list of attributes and the one getting the family. In Enrich, the normalized attributes are embed in the family. It appeared that the normalization of attributes was the common bottleneck.

**The solution**
In the current enrich attribute normalizer we set `created`, `updated`, `structure_version` and `model_type` keys. None of these are used in the PEF, so I split the normalizer in two : `AttributeNormalizer` and `VersionedAttributeNormalizer` (not sure about the naming), the second one decorating the first one. Then use the new one in the family normalizer and in the index route of the attribute rest controller.

For a PEF with 200 attributes :

Request | Before | After
--- | -------- | ------
fetch family | 2,4s | 320ms
fetch attributes | 2,2s | 300ms